feat: split settings and statistics into child routes
Move the settings and statistics sections from internal tab state to
dedicated child routes so the URL is the source of truth.

Each settings/statistics screen now has its own route, with shared
section chrome in layouts. Statistics also restores the last visited
child route when resolving /statistics, while validating persisted
route values before navigation.

Fixes #61.
