--- src/routes/settings/+page.svelte
@@ -1,205 +1,0 @@
-<script lang="ts">
-  import { onMount } from 'svelte';
-  import { tap } from 'rxjs';
-  import SettingsContent from '$lib/components/settings/settings-content.svelte';
-  import SettingsHeader from '$lib/components/settings/settings-header.svelte';
-  import { pxScreen } from '$lib/css-classes';
-  import {
-    addCharactersOnCompletion$,
-    adjustStatisticsAfterIdleTime$,
-    autoBookmark$,
-    autoBookmarkTime$,
-    autoPositionOnResize$,
-    autoReplication$,
-    avoidPageBreak$,
-    cacheStorageData$,
-    confirmClose$,
-    customReadingPointEnabled$,
-    disableWheelNavigation$,
-    enableFontVPAL$,
-    enableReaderWakeLock$,
-    enableTapEdgeToFlip$,
-    enableTextJustification$,
-    enableTextWrapPretty$,
-    enableVerticalFontKerning$,
-    firstDimensionMargin$,
-    fontFamilyGroupOne$,
-    fontFamilyGroupTwo$,
-    fontSize$,
-    furiganaStyle$,
-    hideFurigana$,
-    hideSpoilerImage$,
-    importHTMLFixMode$,
-    lineHeight$,
-    manualBookmark$,
-    keepLocalStatisticsOnDeletion$,
-    openTrackerOnCompletion$,
-    overwriteBookCompletion$,
-    pageColumns$,
-    pauseTrackerOnCustomPointChange$,
-    prioritizeReaderStyles$,
-    replicationSaveBehavior$,
-    restrictImportFixToAnchor$,
-    secondDimensionMaxValue$,
-    selectionToBookmarkEnabled$,
-    showCharacterCounter$,
-    showPercentage$,
-    showFooterChapterCharacterCounter$,
-    showFooterChapterPercentage$,
-    showExternalPlaceholder$,
-    startDayHoursForTracker$,
-    statisticsEnabled$,
-    statisticsMergeMode$,
-    swipeThreshold$,
-    textIndentation$,
-    textMarginMode$,
-    textMarginValue$,
-    theme$,
-    trackerAutoPause$,
-    trackerBackwardSkipThreshold$,
-    trackerForwardSkipThreshold$,
-    trackerAutostartTime$,
-    trackerIdleTime$,
-    trackerPopupDetection$,
-    trackerSkipThresholdAction$,
-    verticalTextOrientation$,
-    viewMode$,
-    writingMode$,
-    readingGoalsMergeMode$,
-    hideSpoilerImageMode$
-  } from '$lib/data/store';
-  import { storage } from '$lib/data/window/navigator/storage';
-  import { formatPageTitle } from '$lib/functions/format-page-title';
-  import { writableSubject } from '$lib/functions/svelte/store';
-  import { reduceToEmptyString } from '$lib/functions/rxjs/reduce-to-empty-string';
-
-  const persistentStorage$ = writableSubject(false);
-  let persistentStorageReactive = false;
-
-  onMount(() => {
-    storage.persisted().then(setPersistentStorage);
-
-    setStorageQuota();
-  });
-
-  let activeSettings = $state('Reader');
-
-  let storageQuota = $state('');
-
-  const setPersistentStorage$ = persistentStorage$.pipe(
-    tap((value) => {
-      if (!persistentStorageReactive) return;
-      if (!value) {
-        setPersistentStorage(true);
-        return;
-      }
-
-      storage.persist().then(setPersistentStorage).finally(setStorageQuota);
-    }),
-    reduceToEmptyString()
-  );
-
-  function setPersistentStorage(value: boolean) {
-    persistentStorageReactive = false;
-    persistentStorage$.next(value);
-    persistentStorageReactive = true;
-  }
-
-  function setStorageQuota() {
-    storage
-      .estimate()
-      .then((storageData) => {
-        const { usage, quota } = storageData;
-
-        if (usage === undefined || quota === undefined) {
-          return;
-        }
-
-        storageQuota = `${Math.round(((usage / quota) * 100 + Number.EPSILON) * 100) / 100} % used`;
-      })
-      .catch(() => {
-        // no-op
-      });
-  }
-</script>
-
-<svelte:head>
-  <title>{formatPageTitle('Settings')}</title>
-</svelte:head>
-
-<div class="elevation-4 fixed inset-x-0 top-0 z-10">
-  <SettingsHeader bind:activeSettings />
-</div>
-
-<div class="{pxScreen} h-full pt-16 xl:pt-14">
-  <div class="max-w-5xl">
-    <SettingsContent
-      {activeSettings}
-      {storageQuota}
-      bind:selectedTheme={$theme$}
-      bind:fontFamilyGroupOne={$fontFamilyGroupOne$}
-      bind:fontFamilyGroupTwo={$fontFamilyGroupTwo$}
-      bind:fontSize={$fontSize$}
-      bind:lineHeight={$lineHeight$}
-      bind:textIndentation={$textIndentation$}
-      bind:textMarginValue={$textMarginValue$}
-      bind:blurImage={$hideSpoilerImage$}
-      bind:blurImageMode={$hideSpoilerImageMode$}
-      bind:hideFurigana={$hideFurigana$}
-      bind:furiganaStyle={$furiganaStyle$}
-      bind:writingMode={$writingMode$}
-      bind:enableFontKerning={$enableVerticalFontKerning$}
-      bind:enableFontVPAL={$enableFontVPAL$}
-      bind:verticalTextOrientation={$verticalTextOrientation$}
-      bind:prioritizeReaderStyles={$prioritizeReaderStyles$}
-      bind:enableTextJustification={$enableTextJustification$}
-      bind:enableTextWrapPretty={$enableTextWrapPretty$}
-      bind:textMarginMode={$textMarginMode$}
-      bind:enableReaderWakeLock={$enableReaderWakeLock$}
-      bind:showCharacterCounter={$showCharacterCounter$}
-      bind:showPercentage={$showPercentage$}
-      bind:showFooterChapterCharacterCounter={$showFooterChapterCharacterCounter$}
-      bind:showFooterChapterPercentage={$showFooterChapterPercentage$}
-      bind:viewMode={$viewMode$}
-      bind:secondDimensionMaxValue={$secondDimensionMaxValue$}
-      bind:firstDimensionMargin={$firstDimensionMargin$}
-      bind:swipeThreshold={$swipeThreshold$}
-      bind:disableWheelNavigation={$disableWheelNavigation$}
-      bind:autoPositionOnResize={$autoPositionOnResize$}
-      bind:avoidPageBreak={$avoidPageBreak$}
-      bind:pauseTrackerOnCustomPointChange={$pauseTrackerOnCustomPointChange$}
-      bind:customReadingPointEnabled={$customReadingPointEnabled$}
-      bind:selectionToBookmarkEnabled={$selectionToBookmarkEnabled$}
-      bind:enableTapEdgeToFlip={$enableTapEdgeToFlip$}
-      bind:pageColumns={$pageColumns$}
-      bind:persistentStorage={$persistentStorage$}
-      bind:confirmClose={$confirmClose$}
-      bind:manualBookmark={$manualBookmark$}
-      bind:autoBookmark={$autoBookmark$}
-      bind:autoBookmarkTime={$autoBookmarkTime$}
-      bind:importHTMLFixMode={$importHTMLFixMode$}
-      bind:restrictImportFixToAnchor={$restrictImportFixToAnchor$}
-      bind:cacheStorageData={$cacheStorageData$}
-      bind:replicationSaveBehavior={$replicationSaveBehavior$}
-      bind:autoReplication={$autoReplication$}
-      bind:showExternalPlaceholder={$showExternalPlaceholder$}
-      bind:keepLocalStatisticsOnDeletion={$keepLocalStatisticsOnDeletion$}
-      bind:overwriteBookCompletion={$overwriteBookCompletion$}
-      bind:startDayHoursForTracker={$startDayHoursForTracker$}
-      bind:statisticsMergeMode={$statisticsMergeMode$}
-      bind:readingGoalsMergeMode={$readingGoalsMergeMode$}
-      bind:statisticsEnabled={$statisticsEnabled$}
-      bind:trackerAutoPause={$trackerAutoPause$}
-      bind:openTrackerOnCompletion={$openTrackerOnCompletion$}
-      bind:addCharactersOnCompletion={$addCharactersOnCompletion$}
-      bind:trackerAutoStartTime={$trackerAutostartTime$}
-      bind:trackerIdleTime={$trackerIdleTime$}
-      bind:trackerForwardSkipThreshold={$trackerForwardSkipThreshold$}
-      bind:trackerBackwardSkipThreshold={$trackerBackwardSkipThreshold$}
-      bind:trackerSkipThresholdAction={$trackerSkipThresholdAction$}
-      bind:trackerPopupDetection={$trackerPopupDetection$}
-      bind:adjustStatisticsAfterIdleTime={$adjustStatisticsAfterIdleTime$}
-    />
-  </div>
-</div>
-{$setPersistentStorage$ ?? ''}
