+++ src/routes/settings/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load(): never {
+  throw redirect(307, '/settings/reader');
+}
